models: many, nested_attributes reject issue

+++ system/models/wstm/cassation.rb
@@ -22,7 +22,7 @@ module Wstm
     scope :by_unit_id, ->(unit_id) {where(unit_id: unit_id)}
 
     accepts_nested_attributes_for :freights,
-      reject_if: ->(attrs){ attrs[:qu].to_i == 0 }
+      reject_if: ->(attrs){ attrs[:qu].to_f == 0 }
 
     class << self
       # @todo
