+++ system/models/wstm/delivery_note.rb
@@ -29,7 +29,7 @@ module Wstm
     scope :by_unit_id, ->(unit_id) {where(unit_id: unit_id)}
 
     accepts_nested_attributes_for :freights,
-      reject_if: ->(attrs){ attrs[:qu].to_i == 0 }
+      reject_if: ->(attrs){ attrs[:qu].to_f == 0 }
 
     class << self
       # @todo
