+++ system/models/wstm/grn.rb
@@ -41,7 +41,7 @@ module Wstm
 
     accepts_nested_attributes_for :dlns
     accepts_nested_attributes_for :freights,
-      reject_if: ->(attrs){ attrs[:qu].to_i == 0 }
+      reject_if: ->(attrs){ attrs[:qu].to_f == 0 }
 
     class << self
       # @todo
